refactor: move CMYK fallback function type into schema
